pyproject.toml
[project]
name = "backend_mooner"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "luansilva", email = "luansilva250807@gmail.com"},
    {name = "luansilva25", email = "luansilva250807@gmail.com"},
    {name = "Gabz047", email = "gabe1612mail@gmail.com"},
]

dependencies = [
    "django>=5.1.2",
    "setuptools>=75.2.0",
    "pillow>=11.0.0",
    "python-magic>=0.4.27",
    "djangorestframework>=3.15.2",
    "passage-identity>=2.1.4",
    "djangorestframework-simplejwt>=5.3.1",
    "google-generativeai>=0.8.3",
    "python-dotenv>=1.0.1",
    "django-cors-headers>=4.5.0",
    "channels>=4.1.0",
    "daphne>=4.1.2",
    "django-filter>=24.3",
    "django-simple-history>=3.7.0",
    "drf-spectacular>=0.27.2",
    "celery>=5.4.0",
    "psycopg2>=2.9.10",
    "psycopg2-binary>=2.9.10",
    "waitress>=3.0.0",
    "cloudinary>=1.41.0",
    "django-cloudinary-storage>=0.3.0",
    "whitenoise>=6.7.0",
    "mercadopago>=2.2.3",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
dev = "python src/manage.py runserver"
createsuperuser = "python src/manage.py createsuperuser"
migrate = "python src/manage.py migrate"
makemigrations = "python src/manage.py makemigrations"
